Change "unsigned int" in delay function to "size_t"

+++ delay.c
@@ -1,9 +1,10 @@
 
 #include "delay.h"
 
+#include <stdio.h>
 #include <time.h>
 
-void delay(unsigned int time) {
+void delay(size_t time) {
 	clock_t end = clock() + time;
 	while (clock() <= end);
 }
